auth/requirements.txt
fastapi==0.78.0
uvicorn==0.16.0
fastapi_utils==0.2.1
python-jose[cryptography]
aerich==0.6.2
tortoise-orm==0.18.0
aerich==0.6.2
pyjwt
redis==4.3.1
passlib==1.7.4
httpx==0.23.0
asyncpg==0.25.0
aioredis==2.0.1
pydantic[dotenv]